reorganize generators/base
