+++ generators/base/generator-base-definitions.mjs
