+++ generators/base/index.mts
@@ -17,4 +17,4 @@
  * limitations under the License.
  */
 
-export { default } from './generator.mjs';
+export { default } from './generator-base-blueprint.mjs';
