+++ generators/server/cleanup-oauth2.mts
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import type BaseGenerator from '../base/generator.mjs';
+import type BaseGenerator from '../base/index.mjs';
 import { type GeneratorDefinition as ServerGeneratorDefinition } from './index.mjs';
 
 type WritingTaskParam = ServerGeneratorDefinition['writingTaskParam'];
